Simplify the connect user code a little
The code in CConnectUserTimer::RunJob() is actually more complicated now, but
the semantics are easier, because we don't cache an iterator between timer runs
and thus the timer doesn't need to be restarted if the user map is changed
(which could accidentally be forgotten).

Sadly, this doesn't work currently. My first user connects, my second one
doesn't. :'(

Signed-off-by: Uli Schlachter <psychon@znc.in>